requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
requests==2.26.0
line-bot-sdk==1.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.22
matplotlib==3.4.2
apscheduler==3.7.0
python-dotenv==0.20.0
pymysql==1.0.2
google-cloud-aiplatform
google-auth==2.11.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0





